* Fixes NB#93291, properly shown email size in details dialogs
pmo-trunk-r7444

+++ src/modest-text-utils.c
@@ -1661,19 +1661,19 @@ modest_text_utils_get_display_size (guint64 size)
 	const guint GB=1024 * MB;
 
 	if (size == 0)
-		return g_strdup_printf (_FM("sfil_li_size_kb"), 0);
+		return g_strdup_printf (_FM("sfil_li_size_kb"), (int) 0);
 	if (0 <= size && size < KB)
-		return g_strdup_printf (_FM("sfil_li_size_1kb_99kb"), 1);
+		return g_strdup_printf (_FM("sfil_li_size_1kb_99kb"), (int) 1);
 	else if (KB <= size && size < 100 * KB)
-		return g_strdup_printf (_FM("sfil_li_size_1kb_99kb"), size / KB);
+		return g_strdup_printf (_FM("sfil_li_size_1kb_99kb"), (int) size / KB);
 	else if (100*KB <= size && size < MB)
 		return g_strdup_printf (_FM("sfil_li_size_100kb_1mb"), (float) size / MB);
 	else if (MB <= size && size < 10*MB)
 		return g_strdup_printf (_FM("sfil_li_size_1mb_10mb"), (float) size / MB);
 	else if (10*MB <= size && size < GB)
-		return g_strdup_printf (_FM("sfil_li_size_10mb_1gb"), size / MB);
+		return g_strdup_printf (_FM("sfil_li_size_10mb_1gb"), (int) size / MB);
 	else
-		return g_strdup_printf (_FM("sfil_li_size_1gb_or_greater"), (float) size / GB);	
+		return g_strdup_printf (_FM("sfil_li_size_1gb_or_greater"), (float) size / GB);
 }
 
 static gchar *
@@ -1690,11 +1690,11 @@ get_email_from_address (const gchar * address)
 		return g_strndup (left_limit + 1, (right_limit - left_limit) - 1);
 }
 
-gchar *      
+gchar *
 modest_text_utils_get_color_string (GdkColor *color)
 {
 	g_return_val_if_fail (color, NULL);
-	
+
 	return g_strdup_printf ("#%x%x%x%x%x%x%x%x%x%x%x%x",
 				(color->red >> 12)   & 0xf, (color->red >> 8)   & 0xf,
 				(color->red >>  4)   & 0xf, (color->red)        & 0xf,
